pyproject.toml
[project]
name = "nano-claw"
version = "0.1.0"
description = "nano-claw Backend"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0",
    "flask-sqlalchemy>=3.1",
    "flask-cors>=4.0",
    "pymysql>=1.1",
    "requests>=2.31",
    "pyyaml>=6.0",
    "duckduckgo-search>=8.0",
    "ddgs>=8.0",
    "beautifulsoup4>=4.12",
    "httpx>=0.25",
    "lxml>=5.0",
    "PyJWT>=2.8",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["backend*"]
